wordcloud codes completed
wordcloud.R contains the codes to open a txt file, calculate word
frequencies, and draw a wordcloud. It could either take all the words
appeared into consideration or the words after screening the most common
words in English. Also, besides wordcloud, it could generate a bar chart
for the most frequently appearing words. trump.txt is a file for testing
purposes, it is one of Trump's congressional speech.